+++ crates/sled/src/materializer.rs
@@ -28,11 +28,11 @@ impl Materializer for BLinkMaterializer {
         }
     }
 
-    fn size_in_bytes(frag: &Frag) -> usize {
+    fn size_in_bytes(frag: &Frag) -> u64 {
         match *frag {
-            Frag::Base(ref node) => std::mem::size_of::<Frag>()
-                .saturating_add(node.size_in_bytes() as usize),
-            _ => std::mem::size_of::<Frag>(),
+            Frag::Base(ref node) => (std::mem::size_of::<Frag>() as u64)
+                .saturating_add(node.size_in_bytes()),
+            _ => std::mem::size_of::<Frag>() as u64,
         }
     }
 }
